+++ server/mock/store.go
@@ -1,6 +1,8 @@
 package mock
 
 import (
+	"io"
+
 	dbm "github.com/tendermint/tendermint/libs/db"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -15,6 +17,26 @@ func (ms multiStore) CacheMultiStore() sdk.CacheMultiStore {
 }
 
 func (ms multiStore) CacheWrap() sdk.CacheWrap {
+	panic("not implemented")
+}
+
+func (ms multiStore) CacheWrapWithTrace(_ io.Writer, _ sdk.TraceContext) sdk.CacheWrap {
+	panic("not implemented")
+}
+
+func (ms multiStore) ResetTraceContext() sdk.MultiStore {
+	panic("not implemented")
+}
+
+func (ms multiStore) TracingEnabled() bool {
+	panic("not implemented")
+}
+
+func (ms multiStore) WithTracingContext(tc sdk.TraceContext) sdk.MultiStore {
+	panic("not implemented")
+}
+
+func (ms multiStore) WithTracer(w io.Writer) sdk.MultiStore {
 	panic("not implemented")
 }
 
@@ -67,6 +89,10 @@ type kvStore struct {
 }
 
 func (kv kvStore) CacheWrap() sdk.CacheWrap {
+	panic("not implemented")
+}
+
+func (kv kvStore) CacheWrapWithTrace(w io.Writer, tc sdk.TraceContext) sdk.CacheWrap {
 	panic("not implemented")
 }