+++ types/store.go
@@ -2,6 +2,7 @@ package types
 
 import (
 	"fmt"
+	"io"
 
 	abci "github.com/tendermint/tendermint/abci/types"
 	cmn "github.com/tendermint/tendermint/libs/common"
@@ -50,6 +51,21 @@ type MultiStore interface { //nolint
 	GetStore(StoreKey) Store
 	GetKVStore(StoreKey) KVStore
 	GetKVStoreWithGas(GasMeter, StoreKey) KVStore
+
+	// TracingEnabled returns if tracing is enabled for the MultiStore.
+	TracingEnabled() bool
+
+	// WithTracer sets the tracer for the MultiStore that the underlying
+	// stores will utilize to trace operations. A MultiStore is returned.
+	WithTracer(w io.Writer) MultiStore
+
+	// WithTracingContext sets the tracing context for a MultiStore. It is
+	// implied that the caller should update the context when necessary between
+	// tracing operations. A MultiStore is returned.
+	WithTracingContext(TraceContext) MultiStore
+
+	// ResetTraceContext resets the current tracing context.
+	ResetTraceContext() MultiStore
 }
 
 // From MultiStore.CacheMultiStore()....
@@ -163,25 +179,27 @@ type KVStoreGetter interface {
 //----------------------------------------
 // CacheWrap
 
-/*
-	CacheWrap() makes the most appropriate cache-wrap.  For example,
-	IAVLStore.CacheWrap() returns a CacheKVStore.
-
-	CacheWrap() should not return a Committer, since Commit() on
-	cache-wraps make no sense.  It can return KVStore, HeapStore,
-	SpaceStore, etc.
-*/
+// CacheWrap makes the most appropriate cache-wrap. For example,
+// IAVLStore.CacheWrap() returns a CacheKVStore. CacheWrap should not return
+// a Committer, since Commit cache-wraps make no sense. It can return KVStore,
+// HeapStore, SpaceStore, etc.
 type CacheWrap interface {
-
 	// Write syncs with the underlying store.
 	Write()
 
 	// CacheWrap recursively wraps again.
 	CacheWrap() CacheWrap
+
+	// CacheWrapWithTrace recursively wraps again with tracing enabled.
+	CacheWrapWithTrace(w io.Writer, tc TraceContext) CacheWrap
 }
 
 type CacheWrapper interface { //nolint
+	// CacheWrap cache wraps.
 	CacheWrap() CacheWrap
+
+	// CacheWrapWithTrace cache wraps with tracing enabled.
+	CacheWrapWithTrace(w io.Writer, tc TraceContext) CacheWrap
 }
 
 //----------------------------------------
@@ -298,3 +316,7 @@ func (getter PrefixStoreGetter) KVStore(ctx Context) KVStore {
 type KVPair cmn.KVPair
 
 //----------------------------------------
+
+// TraceContext contains TraceKVStore context data. It will be written with
+// every trace operation.
+type TraceContext map[string]interface{}